chore: split account into account and account_user
This was done to reduce duplication of account names and associate users
with various accounts. Account could contain rules about how that
account works/ description of that account e.g. savings accounts
shouldn't withdraw regularly

At first, I knew this should be multiple commits/ multiple migrations,
especially in cases when uptime is important but was like meh, not
really needed, wanted to keep things as simple as possible... Then this
happened, much bigger change than I anticipated. Also, how I had it
initially setup meant that a lot of data was actually lost because bill/
register only kept the id of the account and not the id (which didn't
previously exist) of account_user, which in the future should contain
information like card details or description of that account.

I did manually check the database state before and after and looks
correct - checked name and which user the account should be associated
with

Contemplating removing user owns account because in cases of refunds, an
account the user doesn't own is the payer. But I don't know if I should
remove that column altogether because can be beneficial to know which
accounts to display as user's accounts to manage but might also not own
an account but still want to display it e.g. a business tracking an
account managed by an employee to protect against fraud.

GetAccountsOwnedByUser() returning account_user id because this is the
specific account id associated with that user.

previously didn't really want payee and payer as their names are very
similar which can cause confusion but payee and account can cause
confusion as to what account is whereas payer payes the payee. Account
id should all be renamed to payer

For payee, GetAccountUserEntity should still be used because need to
ensure user has an account that is associated with them.

Functionality and tests for adding, editing and removing account and
account_user to follow